+++ MidicalPrescriptionProject/src/main/java/app/core/repositories/PrescriptionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.core.entities.Prescription;
 
-public interface PrescriptionRepository extends JpaRepository<Prescription, Integer> {
+public interface PrescriptionRepository extends JpaRepository<Prescription, Long> {
 
 }
